PC-325 Integration Tests - Basic patient creation abilities
PC-325: Improved verifyEmailNotifications() test by actually checking that there are two emails there.

PC-325 Fix compilation issue, put java 1.8 as a build dependency in the pom file.

PC-325: Adding methods and selectors to allow for manual patient creation. Will continue tomorrow.

PC-375: Adding a test to create a basic patient with phenotypes/genotypes manually. Still working on ensuring test is stable enough for different browser window sizes. For some reason, selenium isn't scrolling some elements to the viewport on Firefox. Need to investigate some more.

PC-325: Expanded create patient test case. Investigating why selenium doesn't automatically scroll. Probably going to do a dpeendency update.

PC-375: Updated dependency of Selenium to 3.1415. Debugged dropdown not actually selecting due to missed click and wrong selector. CreatePatientTest should now work properly. Going to double check.

+++ src/main/java/PageObjects/EmailUIPage.java
@@ -1,31 +1,75 @@
 package PageObjects;
 
+import java.util.List;
+import java.util.Vector;
+
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
 /**
- * This is the webpage of the fake SMTP service's UI. In this case, we are using MockMock.
- * As it provides a web interface for the email inbox.
- * Assumes that the fake SMTP service is running as per the Readme.
+ * This is the webpage of the fake SMTP service's UI. In this case, we are using MockMock. As it provides a web
+ * interface for the email inbox. Assumes that the fake SMTP service is running as per the Readme.
  */
 public class EmailUIPage extends BasePage
 {
-    EmailUIPage(WebDriver aDriver)
+    private final By deleteAllEmailsLink = By.cssSelector(".delete");
+
+    private final By emailStatus = By.cssSelector("div.container:nth-of-type(2) > *:first-child");
+
+    private final By emailRows = By.cssSelector("table > tbody > tr");
+
+    private final By emailTitles = By.cssSelector("table > tbody > tr > td:nth-child(3) > a");
+
+    public EmailUIPage(WebDriver aDriver)
     {
         super(aDriver);
     }
 
-    private final By deleteAllEmailsLink = By.cssSelector(".delete");
-
     /**
-     * Deletes all emails in the inbox.
+     * Deletes all emails, if any, in the inbox.
+     *
      * @return stay on the same page so return the same object.
      */
-    public EmailUIPage deleteAllEmails() {
-        unconditionalWaitNs(5);
-        clickOnElement(deleteAllEmailsLink);
+    public EmailUIPage deleteAllEmails()
+    {
+        if (getNumberOfEmails() > 0) {
+            unconditionalWaitNs(5);
+            clickOnElement(deleteAllEmailsLink);
+        }
+        else {
+            System.out.println("There are no emails to delete!");
+        }
         return this;
     }
 
+    /**
+     * Get a (possibly empty) list of titles of emails from the inbox.
+     * @return a List of Strings which are titles of all the emails.
+     */
+    public List<String> getEmailTitles() {
+        List<String> loTitles = new Vector<String>();
+
+        unconditionalWaitNs(1);
+        if (getNumberOfEmails() > 0) {
+            waitForElementToBePresent(emailTitles);
+            superDriver.findElements(emailTitles).forEach(x -> loTitles.add(x.getText()));
+        }
+        return loTitles;
+    }
+
+    /**
+     * Indicates the number of emails in the inbox. Checks the main text for the overall status first.
+     * @return integer >= 0.
+     */
+    public int getNumberOfEmails() {
+        waitForElementToBePresent(emailStatus);
+        String emailText = superDriver.findElement(emailStatus).getText();
+        if (emailText.contains("You have ")) {
+            return superDriver.findElements(emailRows).size();
+        }
+        else {
+            return 0;
+        }
+    }
 
 }
